# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EmailSecurity API
  version: "1.1"
  description: OpenAPI spec for EmailSecurity backend. Import this into Postman.
servers:
  - url: http://127.0.0.1:8000
    description: Local dev
security:
  - bearerAuth: []
paths:
...
  /emails:
    get:
      summary: List emails
      parameters:
        - in: query
          name: folder
          schema:
            type: string
            enum: [quarantine, blocked, all]
          required: false
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          required: false
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
          required: false
        - in: query
          name: sort
          schema:
            type: string
          required: false
        - in: query
          name: q
          schema:
            type: string
          required: false
        - in: query
          name: date_from
          schema:
            type: string
            format: date-time
          required: false
        - in: query
          name: date_to
          schema:
            type: string
            format: date-time
          required: false
      responses:
        '200':
          description: Paged list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailsPage'
  /emails/{id}:
    get:
      summary: Email detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailDetail'
        '404':
          description: Not found
    delete:
      summary: Delete email (Analyst+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: string
  /emails/{id}/forward:
    post:
      summary: Forward email (Analyst+)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Forwarded placeholder
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    example: forwarded
  /emails/{id}/attachments:
    get:
      summary: List attachments metadata
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Attachments
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttachmentMeta'
  /emails/{id}/attachments/{attId}:
    get:
      summary: Download attachment (placeholder)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: attId
          required: true
          schema:
            type: string
      responses:
        '404':
          description: Content not stored
  /emails/clear:
    post:
      summary: Bulk clear by folder (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClearBody'
      responses:
        '200':
          description: Deleted count
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EmailItem:
      type: object
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        from:
          type: string
        subject:
          type: string
        final_decision:
          type: string
        threat_score:
          type: integer
          nullable: true
    EmailsPage:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmailItem'
    EmailDetail:
      type: object
      properties:
        id:
          type: string
        email:
          type: object
          properties:
            subject:
              type: string
            from:
              type: string
            date:
              type: string
            path:
              type: string
        auth:
          type: object
        attachments:
          type: array
          items:
            type: object
        final_decision:
          type: string
        reasons:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    AttachmentMeta:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        is_malicious:
          type: boolean
    ClearBody:
      type: object
      required: [folder]
      properties:
        folder:
          type: string
          enum: [quarantine, blocked]
...